Remove all references to "example" and "app prototype", instead opting for "gitmad" and "duofit"
