--- app/src/main/java/com/example/appprototype/ui/SettingsScreen.kt
@@ -1,8 +1,0 @@
-package com.example.appprototype.ui
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun SettingsScreen() {
-
-}
