--- app/src/main/java/com/example/appprototype/viewmodels/MessagesViewModel.kt
@@ -1,7 +1,0 @@
-package com.example.appprototype.viewmodels
-
-import androidx.lifecycle.ViewModel
-
-class MessagesViewModel : ViewModel() {
-
-}
