--- app/src/main/java/com/example/appprototype/viewmodels/SettingsViewModel.kt
@@ -1,2 +1,0 @@
-package com.example.appprototype.viewmodels
-
