+++ app/src/main/java/com/gitmad/duofit/models/Profile.kt
@@ -1,0 +1,11 @@
+package com.gitmad.duofit.models
+
+
+data class Profile (
+    val profileID: String = "-1",
+    var name: String = "N/A",
+    var schedule: String = "N/A",
+    var interests: String = "N/A",
+    var mainGym: String = "N/A",
+    var description: String = "N/A",
+    var isFavorite: Boolean = false)
