+++ app/src/main/java/com/gitmad/duofit/ui/SettingsScreen.kt
@@ -1,0 +1,8 @@
+package com.gitmad.duofit.ui
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun SettingsScreen() {
+
+}
