+++ app/src/main/java/com/gitmad/duofit/viewmodels/MessagesViewModel.kt
@@ -1,0 +1,7 @@
+package com.gitmad.duofit.viewmodels
+
+import androidx.lifecycle.ViewModel
+
+class MessagesViewModel : ViewModel() {
+
+}
