+++ app/src/main/java/com/gitmad/duofit/viewmodels/SettingsViewModel.kt
@@ -1,0 +1,2 @@
+package com.gitmad.duofit.viewmodels
+
