+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">AppPrototype</string>
+    <string name="app_name">DuoFit</string>
 </resources>
